gitService/pyproject.toml
[project]
name = "gitservice"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "ivnstd", email = "nasty-iv@bk.ru" }
]
dependencies = [
    "gitpython>=3.1.42",
    "Repo>=0.3.0",
    "requests>=2.31.0",
    "fastapi>=0.109.2",
    "pydantic>=2.6.2",
    "uvicorn>=0.27.1",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "gitservice:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/gitservice"]
